Cryptographic Primitives

Curve Domain Parameters

BinaryECC.CurveDomainParams — Type

CurveDomainParams{D,R}

Represents the elliptic curve domain parameters over $\mathbb{F}_{2^m}$ described in SEC 1 (version 2), 3.1.2.

It contains three fields:

$G$, a generating point, in affine coordinates

$n$, the order of $G$ (i.e. the smallest $n$ satisfying $G \cdot n = \mathcal{O}$)

$h$, the cofactor, $h = \#E(\mathbb{F}_{2^m}) / n$

The other elements of the septuple described in 3.1.2 are accessible through the fields of $G$.

Several standard curves domain parameters (taken from SEC 2, section 3) are available:

SECT163K1

SECT163R1

SECT233K1

SECT233R1

SECT283K1

SECT283R1

SECT409K1

SECT409R1

SECT571K1

SECT571R1

Base.isvalid — Method

isvalid(T::CurveDomainParams{D,R}, t::Int) where {D,R}

Returns true if the curve domain parameters $T$ meet the security level $t$, using the procedure in SEC 1 (version 2) 3.1.2.2.1, and false otherwise.

Note: does not currently perform step 6 (checking that $n$ is prime).

Elliptic Curve Key Pairs

BinaryECC.ECKeyPair — Type

ECKeyPair{D,R}

Represents an elliptic curve key pair (described in SEC 1, version 2, 3.2) with fields $d$ and $Q$ (where $Q = d \cdot G$, and $G$ is the generator of the curve domain paramters used to generate this key pair).

BinaryECC.generate_keypair — Method

generate_keypair(T::CurveDomainParams{D,R}) where {D,R}

Gnerates a new random ECKeyPair associated with T, as described in SEC 1 (version 2) 3.2.1.

Base.isvalid — Method

isvalid(T::CurveDomainParams{D,R}, Q::ECPointAffine{D,R}) where {D,R}

Returns true if $Q$ is a valid public key associated with the curve domain parameters $T$, using the procedure in SEC 1 (version 2) 3.2.2.1, and false otherwise.

ECDSA

BinaryECC.ECDSASignature — Type

ECDSASignature

Represents a signature produced by ECDSA (Elliptic Curve DSA), with the fields $r$ and $s$ (both integers).

BinaryECC.ecdsa_sign — Method

ecdsa_sign(T::CurveDomainParams{D,R}, U::ECKeyPair{D,R}, M::String) where {D,R}

Creates an ECDSASignature using the key pair $U$ (associated with the curve domain parameters $T$) for the message $M$ (a string).

This follows the signing procedure described in SEC 1 (version 2) 4.1.3.

BinaryECC.ecdsa_verify — Method

ecdsa_verify(T::CurveDomainParams{D,R}, Q::ECPointAffine{D,R}, sig::ECDSASignature, M::String) where {D,R}

Returns true if $\textit{sig}$ is valid signature for message $M$ and public key $Q$ (associated with curve domain parameters $T$), following the verifying operation described in SEC 1 (version 2) 4.1.4, and false otherwise.

ECDH

BinaryECC.ecdh_calculate — Method

ecdh_calculate(T::CurveDomainParams{D,R}, dU::PFieldPoint, QV::ECPointAffine{D,R}) where {D,R}

Calculates the shared secret value for entity "U"'s private key ($\textit{dU}$) and entity "V"'s public key ($\textit{QV}$), which are associated with curve domain parameters $T$.

This follows the procedure described in SEC 1 (version 2) 3.3.1.

BinaryECC.ecdh_deployment1 — Method

ecdh_deployment1(T::CurveDomainParams)

Performs the first stage of the ECDH deployment operation (described in SEC 1, version 2, 6.1.2) from the perspective of entity "U".

BinaryECC.ecdh_deployment2 — Method

ecdh_deployment2(T::CurveDomainParams{D,R}, QV::ECPointAffine{D,R}) where {D,R}

Performs the second stage of the ECDH deployment operation (described in SEC 1, version 2, 6.1.2) from the perspective of entity "U", using entity "V"'s public key ($\textit{QV}$).

BinaryECC.ecdh_agreement — Method

ecdh_agreement(T::CurveDomainParams{D,R}, ukey::ECKeyPair{D,R}, QV::ECPointAffine{D,R}) where {D,R}

This performs the ECDH key agreement operation as described in SEC 1 (version 2) 6.1.3.

It is performed from the perspective of entity "U", using their ECKeyPair $\textit{ukey}$ and the public key of entity "V" ($\textit{QV}$).