Binary Field Arithmetic

Types

BinaryECC.BFieldPoint — Type

BFieldPoint{D,R}

Represents a point in the binary field which has order $2^D$ and reduction polynomial

$x^D + x^{r_n} + \cdots + x^{r_0}$

where $R = r_n r_{n-1}\ldots r_1 r_0$ in binary.

Types for points in the standard fields (taken from SEC 2, table 3) are available:

BFieldPoint113

BFieldPoint131

BFieldPoint163

BFieldPoint193

BFieldPoint233

BFieldPoint239

BFieldPoint283

BFieldPoint409

BFieldPoint571

BinaryECC.BFieldPoint — Method

BFieldPoint{D,R}(s::String) where {D,R}

Using the procedure set out in SEC 1 (version 2) 2.3.6, this converts a hex string to a field element.

Arithmetic

General Arithmetic

Base.:== — Method

==(a::BFieldPoint{D,R}, b::BFieldPoint{D,R}) where {D,R}

Returns true if the points $a$ and $b$ from the same field are equal, and false otherwise.

Base.:+ — Method

+(a::BFieldPoint{D,R}, b::BFieldPoint{D,R}) where {D,R}

Returns a new element (of the binary field represented by {D,R}) which is the result of $a+b$.

Base.:- — Method

-(a::BFieldPoint{D,R}, b::BFieldPoint{D,R}) where {D,R}

Returns a new element (of the binary field represented by {D,R}) which is the result of $a-b$.

Base.:* — Method

*(a::BFieldPoint{D,R}, b::BFieldPoint{D,R}) where {D,R}

Returns a new element (of the binary field represented by {D,R}) which is the result of $a \cdot b$.

Base.inv — Method

inv(a::BFieldPoint{D,R}) where {D,R}

Returns a new element $b$ such that $a b ≡ 1 \pmod{f_R(x)}$ (where $f_R(x)$ is the reduction polynomial for the field).

Base.:/ — Method

/(a::BFieldPoint{D,R}, b::BFieldPoint{D,R}) where {D,R}

Returns a new element (of the binary field represented by {D,R}) which is the result of $\frac{a}{b}$.

Base.:^ — Method

^(a::BFieldPoint{D,R}, b::Integer) where {D,R}

Returns a new element (of the binary field represented by {D,R}) which is the result of $a^b$.

Base.sqrt — Method

sqrt(a::BFieldPoint{D,R}) where {D,R}

Returns $b$ such that $b^2 ≡ a \pmod(R)$.

Multiplication

BinaryECC.mult_shiftandadd — Method

mult_shiftandadd(a::BFieldPoint{D,R}, b::BFieldPoint{D,R}) where {D,R}

Returns $a \cdot b$ using the right to left shift and add method.

BinaryECC.mult_threaded — Method

mult_threaded(a::BFieldPoint{D,R}, b::BFieldPoint{D,R}) where {D,R}

Returns $a \cdot b$ using the right to left shift and add method with multithreading.

BinaryECC.mult_ownreduce — Method

mult_ownreduce(a::BFieldPoint{D,R}, b::BFieldPoint{D,R}) where {D,R}

Returns $a \cdot b$ using the right to left shift and add method, without needing to call a reduction function.

BinaryECC.mult_comb_rtl — Method

mult_comb_rtl(a::BFieldPoint{D,R}, b::BFieldPoint{D,R}) where {D,R}

Returns $a \cdot b$ using a right to left comb method (described in Guide to Elliptic Curve Cryptography, algorithm 2.34).

BinaryECC.mult_comb_ltr — Method

mult_comb_ltr(a::BFieldPoint{D,R}, b::BFieldPoint{D,R}) where {D,R}

Returns $a \cdot b$ using a left to right comb method (described in Guide to Elliptic Curve Cryptography, algorithm 2.35).

BinaryECC.mult_comb_window — Method

mult_comb_window(a::BFieldPoint{D,R}, b::BFieldPoint{D,R}, window::Int) where {D,R}

Returns $a \cdot b$ using a left to right comb method windowing (described in Guide to Elliptic Curve Cryptography, algorithm 2.36).

Performs best with a window size of 4.

BinaryECC.square — Method

square(a::BFieldPoint{D,R}) where {D,R}

Returns a new element (of the binary field represented by {D,R}) which is the result of $a^2$.

BinaryECC.square_window — Method

square_window(a::BFieldPoint{D,R}, window::Int) where {D,R}

Returns $a^2$ by inserting a zero between every bit in the original, using the specified window size.

Miscellaneous Arithmetic

Base.iszero — Method

iszero(a::BFieldPoint)

Returns true if $a$ is the zero element of the field represented by D and R, and false otherwise.

Base.zero — Method

zero(::Type{BFieldPoint{D,R}}) where {D,R}

Returns the zero element of the specified field.

Base.isone — Method

isone(a::BFieldPoint)

Returns true if $a$ is equal to one, and false otherwise.

Base.one — Method

one(::Type{BFieldPoint{D,R}}) where {D,R}

Returns element 1 of the specified field.

Base.reduce — Method

reduce(a::BFieldPoint{D,R}) where {D,R}

Returns the least element $b$, such that $a \equiv b \pmod{R}$.

Miscellaneous Functions

BinaryECC.random — Method

random(::Type{BFieldPoint{D,R}}) where {D,R}

Returns a random element of the specified field.

Base.convert — Method

convert(::Type{BigInt}, a::BFieldPoint)

Converts the given field point to a number (of type BigInt), following the procedure set out in SEC 1 (version 2) 2.3.9.