Elliptic Curve Arithmetic

Types

BinaryECC.EC — Type

EC{D,R}

Represents a non-supersingular elliptic curve over the field given by D and R.

Contains fields $a$ and $b$, where:

$y^2 + xy = x^3 + ax^2 + b$

BinaryECC.AbstractECPoint — Type

AbstractECPoint

Abstract type for points on an elliptic curve.

BinaryECC.ECPointAffine — Type

ECPointAffine{D,R} <: AbstractECPoint

Represents a point on an elliptic curve over the field represented by D and R. Contains fields $x$, $y$, and the elliptic field ("ec") that it is on.

$E: y^2 + xy = x^3 + ax^2 + b$

BinaryECC.ECPointJacobian — Type

ECPointJacobian{D,R} <: AbstractECPoint

Represents a point on an elliptic curve over the field represented by D and R. Contains fields $x$, $y$, $z$, and the elliptic field ("ec") that it is on.

$E: y^2 + xyz = x^3 + ax^2z^2 + bz^6$

Each point $(x, y)$ on the curve is represented by a set of equivalent Jacobian points, $\{(\lambda^2 x, \lambda^3 y, \lambda) : \lambda \in K^* \}$ (where $K^*$ is the binary field that the curve is based on).

BinaryECC.ECPointLD — Type

ECPointLD{D,R} <: AbstractECPoint

Represents a point on an elliptic curve over the field represented by D and R. Contains fields $x$, $y$, $z$, and the elliptic field ("ec") that it is on.

$E: y^2 + xyz = x^3z + ax^2z^2 + bz^4$

Each (affine) point $(x, y)$ is represented by a set of Lopez-Dahab points, $\{(\lambda x, \lambda^2 y, \lambda) : \lambda \in K^* \}$ (where $K^*$ is the binary field that the curve is based on).

BinaryECC.ECMismatchException — Type

ECMismatchException <: Exception

Indicates that an operation has been attempted on several points that are not on the same curve.

Curve Functions

Base.:== — Method

==(ec1::EC{D,R}, ec2::EC{D,R}) where {D,R}

Two elliptic curves are equal if they have the same $a$ and $b$ values, and defined over the same field.

Base.repr — Method

repr(ec::EC)

Returns a string representation of an elliptic curve equation, "$y^2 + xy = x^3 + ax^2 + b$".

Point Arithmetic

Base.:== — Method

==(p1::ECPointAffine{D,R}, p2::ECPointAffine{D,R}) where {D,R}

Two points are equal iff they have the same $x$ and $y$ coordinate, and are on the same elliptic curve.

Base.:+ — Method

+(p1::ECPointAffine{D,R}, p2::ECPointAffine{D,R}) where {D,R}

Returns $p_1+p_2$.

If the points are not on the same curve, this will throw an ECMismatchException.

Base.:- — Method

-(p::ECPointAffine) where {D,R}

Returns additive inverse of the given point, $-p$.

Base.:- — Method

-(p1::AbstractECPoint, p2::AbstractECPoint)

Returns $p_1-p_2$.

Base.:* — Method

*(p::ECPointAffine{D,R}, n::Integer) where {D,R}

Returns the result of the scalar multiplication $p \cdot n$, using a double and add method.

Scalar Point Multiplication

BinaryECC.mont_pow_ladder — Function

mont_pow_ladder(p::AbstractECPoint, n::Integer)

Performs $p \cdot n$ with a fixed sequence of curve and field operations. More resistant to timing attacks than the standard double and add algorithm.

mont_pow_ladder(p::ECPointAffine{D,R}, n::Integer) where {D,R}

Returns $p \cdot n$.

More resistant to timing attacks than the standard double and add algorithm.

Fast Multiplication on Elliptic Curves over $GF(2^m)$ without Precomputation, Algorithm 2A: Montgomery Scalar Multiplication.

BinaryECC.naf_mult — Method

naf_mult(p::ECPointAffine{D,R}, n::Integer) where {D,R}

Returns $p \cdot n$.

Uses the binary NAF multiplication method described in Guide to Elliptic Curve Cryptography, algorithm 3.31.

Additive Identity

Base.iszero — Method

iszero(p::ECPointAffine)

Returns true if $p = \mathcal{O}$, i.e it is the point at infinity.

Base.zero — Method

zero(::Type{ECPointAffine{D,R}}, ec::EC{D,R}) where {D,R}

Returns an object representing the point at infinity on the given curve.

Base.zero — Method

zero(::Type{ECPointAffine, ec::EC{D,R}) where {D,R}

Returns an object representing the point at infinity on the given curve.

Base.iszero — Method

iszero(p::ECPointJacobian)

Returns true if $p = \mathcal{O}$, i.e it is the point at infinity.

Base.zero — Method

zero(::Type{ECPointJacobian{D,R}}, ec::EC{D,R}) where {D,R}

Returns an object representing the point at infinity on the given curve.

Base.zero — Method

zero(::Type{ECPointJacobian}, ec::EC{D,R}) where {D,R}

Returns an object representing the point at infinity on the given curve.

Base.iszero — Method

iszero(p::ECPointLD)

Returns true if $p = \mathcal{O}$, i.e it is the point at infinity.

Base.zero — Method

zero(::Type{ECPointLD{D,R}}, ec::EC{D,R}) where {D,R}

Returns an object representing the point at infinity on the given curve.

Base.zero — Method

zero(::Type{ECPointLD}, ec::EC{D,R}) where {D,R}

Returns an object representing the point at infinity on the given curve.

Mixed Representation Functions

Base.:+ — Method

+(::Type{ECPointJacobian{D,R}}, p1::ECPointLD{D,R}, p2::ECPointJacobian{D,R}) where {D,R}

Returns $p_1 +p_2$ in Jacobian coordinates.

Base.:+ — Method

+(::Type{ECPointJacobian{D,R}}, p1::ECPointJacobian{D,R}, p2::ECPointLD{D,R}) where {D,R}

Returns $p_1 +p_2$ in Jacobian coordinates.

Base.:+ — Method

+(::Type{ECPointLD{D,R}}, p1::ECPointLD{D,R}, p2::ECPointJacobian{D,R}) where {D,R}

Returns $p_1 +p_2$ in Lopez-Dahab coordinates.

Base.:+ — Method

+(::Type{ECPointLD{D,R}}, p1::ECPointJacobian{D,R}, p2::ECPointLD{D,R}) where {D,R}

Returns $p_1 +p_2$ in Lopez-Dahab coordinates.

Base.convert — Method

convert(::Type{ECPointAffine}, p::ECPointJacobian{D,R}) where {D,R}

Converts a point from Jacobian coordinates to affine coordinates.

Base.convert — Method

convert(::Type{ECPointAffine}, p::ECPointLD{D,R}) where {D,R}

Converts a point from Lopez-Dahab coordinates to affine coordinates.

Base.convert — Method

convert(::Type{ECPointJacobian}, p::ECPointAffine{D,R}) where {D,R}

Converts a point from affine coordinates to Jacobian coordinates.

Base.convert — Method

convert(::Type{ECPointJacobian}, p::ECPointLD{D,R}) where {D,R}

Converts a point from Lopez-Dahab coordinates to Jacobian coordinates.

Base.convert — Method

convert(::Type{ECPointLD}, p::ECPointAffine{D,R}) where {D,R}

Converts a point from affine coordinates to Lopez-Dahab coordinates.

Base.convert — Method

convert(::Type{ECPointLD}, p::ECPointJacobian{D,R}) where {D,R}

Converts a point from Jacobian coordinates to Lopez-Dahab coordinates.

Miscellaneous Point Functions

BinaryECC.ECPointAffine — Method

ECPointAffine(s::String, ec::EC{D,R}) where {D,R}

Convert a hex string to a point on the given elliptic curve using the procedure in SEC 2 (version 2), section 2.3.4.

BinaryECC.ECPointAffine — Method

ECPointAffine(s::String, ec::EC{D,R}) where {D,R}

Convert a hex string to a point on the given elliptic curve using the procedure in SEC 2 (version 2), section 2.3.4.

Base.isvalid — Method

isvalid(p::ECPointAffine)

Returns true if $p$ is a point on the elliptic curve that it is associated with.

Base.repr — Method

repr(p::ECPointAffine)

Returns a string representation of an elliptic curve point, "$(x, y)$".

Base.repr — Method

repr(p::ECPointJacobian)

Returns a string representation of an elliptic curve point, "$(x, y, z)$".

Base.repr — Method

repr(p::ECPointLD)

Returns a string representation of an elliptic curve point, "$(x, y, z)$".