Binary Field Arithmetic

Types

BinaryECC.FieldPoint — Type

FieldPoint{D,R}

Represents a point in the binary field which has order $2^D$ and reduction polynomial

$x^D + x^{r_n} + \cdots + x^{r_0}$

where $R = r_n r_{n-1}\ldots r_1 r_0$ in binary.

Types for points in the standard fields (taken from SEC 2, table 3) are available:

FieldPoint113

FieldPoint131

FieldPoint163

FieldPoint193

FieldPoint233

FieldPoint239

FieldPoint283

FieldPoint409

FieldPoint571

BinaryECC.FieldPoint — Method

FieldPoint{D,R}(s::String) where {D,R}

Using the procedure set out in SEC 1 (version 2) 2.3.6, this converts a hex string to a field element.

Arithmetic

Base.:== — Method

==(a::FieldPoint{D,R}, b::FieldPoint{D,R}) where {D,R}

Returns true if the points $a$ and $b$ from the same field are equal, and false otherwise.

Base.:+ — Method

+(a::FieldPoint{D,R}, b::FieldPoint{D,R}) where {D,R}

Returns a new element (of the binary field represented by {D,R}) which is the result of $a+b$.

Base.:- — Method

-(a::FieldPoint{D,R}, b::FieldPoint{D,R}) where {D,R}

Returns a new element (of the binary field represented by {D,R}) which is the result of $a-b$.

Base.:- — Method

-(a::FieldPoint{D,R}, b::FieldPoint{D,R}) where {D,R}

Returns a new element (of the binary field represented by {D,R}) which is the result of $a-b$.

-(a::FieldPoint{D,R}) where {D,R}

Returns a new element (of the binary field represented by {D,R}) which is the result of $-a$.

-(p::ECPointAffine) where {D,R}

Returns additive inverse of the given point, $-p$.

Base.:* — Method

*(a::FieldPoint{D,R}, b::FieldPoint{D,R}) where {D,R}

Returns a new element (of the binary field represented by {D,R}) which is the result of $a \cdot b$.

Base.inv — Method

inv(a::FieldPoint{D,R}) where {D,R}

Returns a new element $b$ such that $a b ≡ 1 \pmod{f_R(x)}$ (where $f_R(x)$ is the reduction polynomial for the field).

Base.:/ — Method

/(a::FieldPoint{D,R}, b::FieldPoint{D,R}) where {D,R}

Returns a new element (of the binary field represented by {D,R}) which is the result of $\frac{a}{b}$.

Base.:^ — Method

^(a::FieldPoint{D,R}, b::Integer) where {D,R}

Returns a new element (of the binary field represented by {D,R}) which is the result of $a^b$.

Base.sqrt — Method

sqrt(a::FieldPoint{D,R}) where {D,R}

Returns $b$ such that $b^2 ≡ a \pmod(R)$.

Base.iszero — Method

iszero(a::FieldPoint)

Returns true if $a$ is the zero element of the field represented by D and R, and false otherwise.

Base.zero — Method

zero(::Type{FieldPoint{D,R}}) where {D,R}

Returns the zero element of the specified field.

Base.isone — Method

isone(a::FieldPoint)

Returns true if $a$ is equal to one, and false otherwise.

Base.one — Method

one(::Type{FieldPoint{D,R}}) where {D,R}

Returns element 1 of the specified field.

Miscellaneous Functions

BinaryECC.random — Method

random(::Type{FieldPoint{D,R}}) where {D,R}

Returns a random element of the specified field.

Base.convert — Method

convert(::Type{BigInt}, a::FieldPoint)

Converts the given field point to a number (of type BigInt), following the procedure set out in SEC 1 (version 2) 2.3.9.