Binary Field Arithmetic

Types

BinaryECC.FieldPoint — Type

FieldPoint{D,R}

Represents a point in the binary field which has order $2^D$ and reduction polynomial

$x^D + x^{r_n} + \cdots + x^{r_0}$

where $R = r_n r_{n-1}\ldots r_1 r_0$ in binary.

Types for points in the standard fields (taken from SEC 2, table 3) are available:

FieldPoint113

FieldPoint131

FieldPoint163

FieldPoint193

FieldPoint233

FieldPoint239

FieldPoint283

FieldPoint409

FieldPoint571

BinaryECC.FieldPoint — Method

FieldPoint{D,R}(s::String) where {D,R}

Using the procedure set out in SEC 1 (version 2) 2.3.6, this converts a hex string to a field element.

Arithmetic

General Arithmetic

Base.:== — Method

==(a::FieldPoint{D,R}, b::FieldPoint{D,R}) where {D,R}

Returns true if the points $a$ and $b$ from the same field are equal, and false otherwise.

Base.:+ — Method

+(a::FieldPoint{D,R}, b::FieldPoint{D,R}) where {D,R}

Returns a new element (of the binary field represented by {D,R}) which is the result of $a+b$.

Base.:- — Method

-(a::FieldPoint{D,R}, b::FieldPoint{D,R}) where {D,R}

Returns a new element (of the binary field represented by {D,R}) which is the result of $a-b$.

Base.:* — Method

*(a::FieldPoint{D,R}, b::FieldPoint{D,R}) where {D,R}

Returns a new element (of the binary field represented by {D,R}) which is the result of $a \cdot b$.

Base.inv — Method

inv(a::FieldPoint{D,R}) where {D,R}

Returns a new element $b$ such that $a b ≡ 1 \pmod{f_R(x)}$ (where $f_R(x)$ is the reduction polynomial for the field).

Base.:/ — Method

/(a::FieldPoint{D,R}, b::FieldPoint{D,R}) where {D,R}

Returns a new element (of the binary field represented by {D,R}) which is the result of $\frac{a}{b}$.

Base.:^ — Method

^(a::FieldPoint{D,R}, b::Integer) where {D,R}

Returns a new element (of the binary field represented by {D,R}) which is the result of $a^b$.

Base.sqrt — Method

sqrt(a::FieldPoint{D,R}) where {D,R}

Returns $b$ such that $b^2 ≡ a \pmod(R)$.

Multiplication

BinaryECC.right_to_left_mult — Method

right_to_left_mult(a::FieldPoint{D,R}, b::FieldPoint{D,R}) where {D,R}

Returns $a \cdot b$ using the right to left shift and add method.

BinaryECC.threads_mult — Method

threads_mult(a::FieldPoint{D,R}, b::FieldPoint{D,R}) where {D,R}

Returns $a \cdot b$ using the right to left shift and add method with multithreading.

BinaryECC.noreduce_mult — Method

noreduce_mult(a::FieldPoint{D,R}, b::FieldPoint{D,R}) where {D,R}

Returns $a \cdot b$ using the right to left shift and add method, without needing to call a reduction function.

BinaryECC.right_to_left_comb_mult — Method

right_to_left_comb_mult(a::FieldPoint{D,R}, b::FieldPoint{D,R}) where {D,R}

Returns $a \cdot b$ using a right to left comb method (described in Guide to Elliptic Curve Cryptography, algorithm 2.34).

BinaryECC.left_to_right_comb_mult — Method

left_to_right_comb_mult(a::FieldPoint{D,R}, b::FieldPoint{D,R}) where {D,R}

Returns $a \cdot b$ using a left to right comb method (described in Guide to Elliptic Curve Cryptography, algorithm 2.35).

BinaryECC.window_comb_mult — Method

window_comb_mult(a::FieldPoint{D,R}, b::FieldPoint{D,R}, window::Int) where {D,R}

Returns $a \cdot b$ using a left to right comb method windowing (described in Guide to Elliptic Curve Cryptography, algorithm 2.36).

Performs best with a window size of 4.

BinaryECC.square — Method

square(a::FieldPoint{D,R}) where {D,R}

Returns a new element (of the binary field represented by {D,R}) which is the result of $a^2$.

BinaryECC.window_square — Method

window_square(a::FieldPoint{D,R}, window::Int) where {D,R}

Returns $a^2$ by inserting a zero between every bit in the original, using the specified window size.

Miscellaneous Arithmetic

Base.iszero — Method

iszero(a::FieldPoint)

Returns true if $a$ is the zero element of the field represented by D and R, and false otherwise.

Base.zero — Method

zero(::Type{FieldPoint{D,R}}) where {D,R}

Returns the zero element of the specified field.

Base.isone — Method

isone(a::FieldPoint)

Returns true if $a$ is equal to one, and false otherwise.

Base.one — Method

one(::Type{FieldPoint{D,R}}) where {D,R}

Returns element 1 of the specified field.

Base.reduce — Method

reduce(a::FieldPoint{D,R}) where {D,R}

Returns the least element $b$, such that $a \equiv b \pmod{R}$.

Miscellaneous Functions

BinaryECC.random — Method

random(::Type{FieldPoint{D,R}}) where {D,R}

Returns a random element of the specified field.

Base.convert — Method

convert(::Type{BigInt}, a::FieldPoint)

Converts the given field point to a number (of type BigInt), following the procedure set out in SEC 1 (version 2) 2.3.9.